requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.4.1
greenlet==3.0.1
idna==3.6
iniconfig==2.0.0
multidict==6.0.4
packaging==23.2
playwright==1.40.0
pluggy==1.3.0
pyee==11.0.1
pytest==7.4.3
pytest-base-url==2.0.0
pytest-playwright==0.4.3
python-dotenv==1.0.0
python-slugify==8.0.1
requests==2.31.0
text-unidecode==1.3
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
